+++ src/Resources/contao/languages/de/tl_linktracker.php
@@ -14,7 +14,7 @@ $GLOBALS['TL_LANG']['tl_linktracker']['id'] = array('ID', 'ID des Datensatzes');
 
 $GLOBALS['TL_LANG']['tl_linktracker']['title_legend'] = 'Titel und URL';
 $GLOBALS['TL_LANG']['tl_linktracker']['title'] = array('Titel', 'Name/Bezeichnung der URL');
-$GLOBALS['TL_LANG']['tl_linktracker']['url'] = array('URL', 'URL');
+$GLOBALS['TL_LANG']['tl_linktracker']['url'] = array('URL', 'URL. Leerlassen, wenn nicht weitergeleitet werden soll.');
 $GLOBALS['TL_LANG']['tl_linktracker']['hits'] = array('Aufrufe', 'Aufrufe');
 
 $GLOBALS['TL_LANG']['tl_linktracker']['publish_legend'] = 'Veröffentlichung';